[rename] 네이버지도 컴포넌트 파일 이름 변경 #37
